[hotfix][table-runtime-blink] Rename StateTtlConfigUtil to StateConfigUtil
